Database connection moved to main function
