+++ testDbConnection.go
@@ -1,39 +1,21 @@
 package main
 
 import (
-	"context"
-	"log"
+	"fmt"
 	"net/http"
-	"os"
 
 	"github.com/TDDD96-G11-BLHub/dbman/db"
 	"github.com/gin-gonic/gin"
-	"github.com/joho/godotenv"
 	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-func testDbConnection(c *gin.Context) {
+func testDbConnection(client *mongo.Client) gin.HandlerFunc {
+	fn := func(c *gin.Context) {
+		fmt.Print("testDbConnection")
+		db.TestConnection(client)
 
-	if err := godotenv.Load(); err != nil {
-		log.Println("No .env file found")
-	}
-	uri := os.Getenv("MONGODB_URI")
-	if uri == "" {
-		log.Fatal("You must set your 'MONGODB_URI' environment variable. See\n\t https://www.mongodb.com/docs/drivers/go/current/usage-examples/#environment-variable")
+		c.JSON(http.StatusOK, gin.H{})
 	}
 
-	client, err := mongo.Connect(context.TODO(), options.Client().ApplyURI(uri))
-	if err != nil {
-		panic(err)
-	}
-	defer func() {
-		if err := client.Disconnect(context.TODO()); err != nil {
-			panic(err)
-		}
-	}()
-
-	db.TestConnection(client)
-
-	c.JSON(http.StatusOK, gin.H{})
+	return gin.HandlerFunc(fn)
 }
